+++ CHANGES.rst
@@ -15,9 +15,9 @@ There's a frood who really knows where his towel is.
   Check developers documentation for more information on how to use them (closes `#504`_).
   [rodfersou]
 
-- Fix performance issues when searching for recent items in the content chooser.
+- Fix performance issues in the content chooser.
   Searches only start after the third character and will be delayed by 500ms to wait for the user to stop typing.
-  The scope of the search was also changed: instead of searching on SearchableText index (full text), we search on Title index only (fixes `#493`_).
+  The scope of the search was also changed: instead of searching on SearchableText index (full text), we search on Title index only (fixes `#493`_ and `#494`_).
   [rodfersou]
 
 - Fix the indexing of RichText tiles by manually transform the text/html to text/plain.
@@ -627,5 +627,6 @@ There's a frood who really knows where his towel is.
 .. _`#473`: https://github.com/collective/collective.cover/issues/473
 .. _`#476`: https://github.com/collective/collective.cover/issues/476
 .. _`#493`: https://github.com/collective/collective.cover/issues/493
+.. _`#494`: https://github.com/collective/collective.cover/issues/494
 .. _`#504`: https://github.com/collective/collective.cover/issues/504
 .. _`PloneFormGen`: https://pypi.python.org/pypi/Products.PloneFormGen